+ add members to RegisterResponseDto + rename MainController to RegisterController + remove everything redundant from RegisterController
