+++ src/main/java/com/example/kkbackend/dtos/RegisterResponseDto.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 @Builder
 public record RegisterResponseDto(int membersCount, String message, boolean isAlreadyRegistered,
-                                  List<TelegramMessageDto> messages) {
+                                  List<TelegramMessageDto> messages, List<MemberDto> members) {
 }
